Fetch best account name, issuer where available
For some accounts, originalName and name coexist; for some, only name
exists, and for some, only originalName exists. When both are available,
I usually see that one contains the issuer and the other doesn't (e.g.
originalName = Issuer:user@example.com, name = user@example.com), so
here we try to pick the best one available.

Most accounts come with an issuer as part of the name, but in some cases
it's not available - e.g. name = user@example.com, issuer = None,
account_type = example. In this case, we can guess the issuer from the
account type and help fill in extra information here.

Additionally, I checked the full list of keys for all my account types
and filled in the known list of keys here. If a user has an account with
a field outside this list, hopefully this will provide a clearer warning
to them that the final TOTP code may be invalid!